refactor: refactored contentscript and explainer comments added on top of the files
