+++ src/utils/storage.ts
@@ -1,3 +1,6 @@
+/**
+ * file that works with browser storage apis to get and set user locations and other user preferences
+ */
 export interface SyncStorage {
   userLocations?: UserLocationItems;
   isExtensionEnabled?: boolean;
@@ -10,18 +13,18 @@ export interface UserLocationItem {
 
 export type SyncStorageKeys = keyof SyncStorage;
 
-export function setUserLocationsInStorage(userLocations: UserLocationItems): Promise<void> {
+export function setUserLocationsInStorage(
+  userLocations: UserLocationItems
+): Promise<void> {
   const vals: SyncStorage = {
     userLocations
   };
-  
 
   return new Promise((resolve) => {
     chrome.storage.local.set(vals, () => {
       resolve();
     });
   });
-  
 }
 
 export function getUserLocationsInStorage(): Promise<UserLocationItems> {
@@ -32,21 +35,23 @@ export function getUserLocationsInStorage(): Promise<UserLocationItems> {
     });
   });
 }
-export function setIsExtensionEnabledInStorage(isExtensionEnabled: boolean): Promise<void>{
+export function setIsExtensionEnabledInStorage(
+  isExtensionEnabled: boolean
+): Promise<void> {
   const vals: SyncStorage = {
     isExtensionEnabled
   };
   return new Promise((resolve) => {
     chrome.storage.local.set(vals, () => {
       resolve();
-    })
-  })
+    });
+  });
 }
-export function getIsExtensionEnabledInStorage(): Promise<boolean>{
+export function getIsExtensionEnabledInStorage(): Promise<boolean> {
   const key: SyncStorageKeys[] = ['isExtensionEnabled'];
   return new Promise((resolve) => {
     chrome.storage.local.get(key, (res: SyncStorage) => {
       resolve(res.isExtensionEnabled ?? true);
-    })
-  })
+    });
+  });
 }
